+++ examples/data-sources/domain-forwarding.tf
@@ -1,0 +1,4 @@
+data "godaddy_domain_forwarding" "blog" {
+  fqdn         = "blog.example.com"
+  include_subs = true
+}