+++ examples/data-sources/domain.tf
@@ -1,3 +1,6 @@
 data "godaddy_domain" "example" {
-  domain = "example.com"
+  domain                        = "example.com"
+  include_actions               = true
+  include_dnssec_records        = true
+  include_registry_status_codes = true
 }