+++ examples/data-sources/domains.tf
@@ -1,0 +1,4 @@
+data "godaddy_domains" "all" {
+  limit    = 25
+  includes = ["nameServers"]
+}