+++ examples/resources/domain-settings.tf
@@ -1,0 +1,5 @@
+resource "godaddy_domain_settings" "example" {
+  domain     = "example.com"
+  locked     = true
+  renew_auto = true
+}